requirements.txt
docopt~=0.6.2  # Minor patches avoid breaking changes even for 0.y.z releases
pyYaml~=6.0.1
requests~=2.26.0
progress~=1.6
deepmerge~=0.3.0  # Minor patches avoid breaking changes even for 0.y.z releases
ibisclient~=1.3
ucam-identitylib~=0.1.7  # Minor patches avoid breaking changes even for 0.y.z releases
tenacity~=8.0.1
